backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cssselect==1.4.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.133.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.24.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jieba3k==0.35.1
jiter==0.13.0
joblib==1.5.3
lxml==6.0.2
newspaper3k==0.2.8
nltk==3.9.3
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
requests-file==3.0.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
tinysegmenter==0.3
tldextract==5.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
